moved endpoint publish-products to new SiteContentController
